Move RuntimeThread to Internal.Runtime.Augments
