+++ src/System.Private.CoreLib/src/Internal/Runtime/Augments/RuntimeThread.Unix.cs
@@ -2,10 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Diagnostics;
-using System.Runtime;
+using System.Threading;
 
-namespace System.Threading
+namespace Internal.Runtime.Augments
 {
     public sealed partial class RuntimeThread
     {
