+++ src/System.Private.CoreLib/src/Internal/Runtime/Augments/RuntimeThread.cs
@@ -2,11 +2,13 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Diagnostics;
 using System.Runtime;
+using System.Threading;
 using Microsoft.Win32.SafeHandles;
 
-namespace System.Threading
+namespace Internal.Runtime.Augments
 {
     public sealed partial class RuntimeThread
     {
