+++ src/System.Private.CoreLib/src/Internal/Runtime/CompilerHelpers/StartupCode/ThreadingHelpers.cs
@@ -8,6 +8,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading;
+using Internal.Runtime.Augments;
 
 namespace Internal.Runtime.CompilerHelpers
 {
