+++ src/System.Private.CoreLib/src/System/GC.cs
@@ -17,6 +17,7 @@ using System.Runtime.CompilerServices;
 using System.Diagnostics;
 using System.Diagnostics.Contracts;
 using System.Security;
+using Internal.Runtime.Augments;
 
 namespace System
 {
