+++ src/System.Private.CoreLib/src/System/Threading/FirstLevelSpinWaiter.cs
@@ -4,7 +4,7 @@
 
 using System.Diagnostics;
 using System.Runtime;
-using System.Runtime.CompilerServices;
+using Internal.Runtime.Augments;
 
 namespace System.Threading
 {
