+++ src/System.Private.CoreLib/src/System/Threading/LowLevelMonitor.Unix.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Diagnostics;
+using Internal.Runtime.Augments;
 
 namespace System.Threading
 {
