+++ src/System.Private.CoreLib/src/System/Threading/LowLevelThread.cs
@@ -10,6 +10,7 @@
 
 using System.Diagnostics;
 using System.Runtime;
+using Internal.Runtime.Augments;
 
 #pragma warning disable 0420
