+++ src/System.Private.CoreLib/src/System/Threading/ReaderWriterLockSlim.cs
@@ -3,8 +3,8 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Diagnostics; // for TraceInformation
-using System.Threading;
 using System.Runtime.CompilerServices;
+using Internal.Runtime.Augments;
 
 // TODO: Delete this and use the implementation from CoreFX
 
@@ -1058,7 +1058,7 @@ namespace System.Threading
             {
                 if (i < LockSpinCount && pc > 1)
                 {
-                    System.Threading.SpinWait.Spin(LockSpinCycles * (i + 1)); // Wait a few dozen instructions to let another processor release lock.
+                    RuntimeThread.SpinWait(LockSpinCycles * (i + 1)); // Wait a few dozen instructions to let another processor release lock.
                 }
                 else if (i < (LockSpinCount + LockSleep0Count))
                 {
@@ -1089,7 +1089,7 @@ namespace System.Threading
             //Exponential backoff
             if ((SpinCount < 5) && (Environment.ProcessorCount > 1))
             {
-                System.Threading.SpinWait.Spin(LockSpinCycles * SpinCount);
+                RuntimeThread.SpinWait(LockSpinCycles * SpinCount);
             }
             else if (SpinCount < MaxSpinCount - 3)
             {
