+++ src/System.Private.CoreLib/src/System/Threading/SpinWait.cs
@@ -10,6 +10,7 @@
 
 using System.Diagnostics;
 using System.Runtime;
+using Internal.Runtime.Augments;
 
 namespace System.Threading
 {
