+++ src/System.Private.CoreLib/src/System/Threading/WaitHandle.cs
@@ -11,12 +11,12 @@
 **
 =============================================================================*/
 
-using Microsoft.Win32.SafeHandles;
 using System.Diagnostics;
-using System.Diagnostics.Contracts;
-using System.IO;
-using System.Runtime.InteropServices;
 using System.Diagnostics.CodeAnalysis;
+using System.Diagnostics.Contracts;
+using System.Runtime.InteropServices;
+using Internal.Runtime.Augments;
+using Microsoft.Win32.SafeHandles;
 
 namespace System.Threading
 {
