+++ src/System.Private.CoreLib/src/System/Threading/WaitSubsystem.ThreadWaitInfo.Unix.cs
@@ -4,6 +4,7 @@
 
 using System.Diagnostics;
 using System.Runtime;
+using Internal.Runtime.Augments;
 
 namespace System.Threading
 {
