+++ src/System.Private.Interop/src/System/Runtime/InteropServices/HandleCollector.cs
@@ -4,6 +4,7 @@
 
 using System;
 using System.Threading;
+using Internal.Runtime.Augments;
 
 namespace System.Runtime.InteropServices
 {
